requirements.txt
Flask==3.1.0
flask-cors==5.0.0
supabase==2.13.0
python-dotenv==1.0.1
xgboost==2.1.3
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.3
joblib==1.4.2
